Basic refactoring to get all tests to pass
