+++ lib/transitland_client/entities/route.rb
@@ -1,0 +1,4 @@
+module TransitlandClient
+  class Route < Entity
+  end
+end
