+++ lib/transitland_client/entities/route_stop_pattern.rb
@@ -1,0 +1,4 @@
+module TransitlandClient
+  class RouteStopPattern < Entity
+  end
+end
