+++ spec/lib/inputs/bounding_box_spec.rb
