+++ spec/lib/inputs/onestop_id_spec.rb
